Democratized-RPA-staging_files/Democratized-RPA-staging_files/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
clipboard==0.0.4
Django==3.1.2
django-bootstrap-form==3.4
django-pandas==0.6.2
djangorestframework==3.12.1
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.5
pdfrw==0.4
Pillow==8.0.1
psycopg2==2.8.6
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python-anticaptcha==0.7.1
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tika==1.24
urllib3==1.26.2
xlrd==2.0.1
XlsxWriter==1.3.6
